My original idea of filling out /etc/environment didn't work.
This approach seems a bit more robust, if not still a bit ugly.  It adds
a sage-entrypoint which, when used as the ENTRYPOINT in the Dockerfile,
sets all the default environment variables for the 'sage' user.

Also changed things slightly so that the post_install.sh script is not
run as root by default.